chore(structure): reorganize files to implement Atomic Design principles
🏗️ Restructured project files based on the Atomic Design methodology.
🔄 Moved components into 'atoms', 'molecules', 'organisms', and 'templates' directories.
📁 Created new directories for 'pages' and 'layouts' to better organize higher-level structures.
