+++ e-commerce/src/app/components/Footer/Fotter.styles.ts
