+++ e-commerce/src/app/page.tsx
@@ -1,9 +1,9 @@
 "use client";
 import React from "react";
 import LoginPage from "./login/page.tsx";
-import { Header } from "./components/header/header.tsx";
-import { FooterPage } from "./components/footer/footer.tsx";
-import HomePage from "./home/page.tsx";
+import { Header } from "./components/Header/Header.tsx";
+import { FooterPage } from "./components/Footer/Footer.tsx";
+import HomePage from "./Home/page.tsx";
 
 export default function Home() {
   return (
